refactor(rust): split oversized modules
