+++ src/openhuman/config/schema/load.rs
@@ -647,101 +647,15 @@ fn encrypt_config_secrets(config: &mut Config) -> Result<()> {
     Ok(())
 }
 
-const ACTIVE_USER_STATE_FILE: &str = "active_user.toml";
+#[path = "load_user_state.rs"]
+mod load_user_state;
+#[cfg(test)]
+pub(crate) use load_user_state::ACTIVE_USER_STATE_FILE;
+pub use load_user_state::{
+    clear_active_user, pre_login_user_dir, read_active_user_id, user_openhuman_dir,
+    write_active_user_id, PRE_LOGIN_USER_ID,
+};
 
-#[derive(Debug, Serialize, Deserialize)]
-struct ActiveUserState {
-    user_id: String,
-}
-
-/// Reads the active user id from `{default_openhuman_dir}/active_user.toml`.
-/// Returns `None` when the file does not exist, is empty, or cannot be parsed.
-pub fn read_active_user_id(default_openhuman_dir: &Path) -> Option<String> {
-    let path = default_openhuman_dir.join(ACTIVE_USER_STATE_FILE);
-    let contents = std::fs::read_to_string(&path).ok()?;
-    let state: ActiveUserState = toml::from_str(&contents).ok()?;
-    let id = state.user_id.trim().to_string();
-    if id.is_empty() {
-        None
-    } else {
-        Some(id)
-    }
-}
-
-/// Writes the active user id to `{default_openhuman_dir}/active_user.toml`.
-pub fn write_active_user_id(default_openhuman_dir: &Path, user_id: &str) -> Result<()> {
-    let path = default_openhuman_dir.join(ACTIVE_USER_STATE_FILE);
-    let state = ActiveUserState {
-        user_id: user_id.to_string(),
-    };
-    let toml_str = toml::to_string_pretty(&state).context("serialize active_user.toml")?;
-    std::fs::write(&path, toml_str)
-        .with_context(|| format!("Failed to write active user state: {}", path.display()))?;
-    tracing::debug!(user_id = %user_id, path = %path.display(), "active user written");
-    Ok(())
-}
-
-/// Removes the active user marker.  After this, the next config load will
-/// use the default (unauthenticated) openhuman directory.
-pub fn clear_active_user(default_openhuman_dir: &Path) -> Result<()> {
-    let path = default_openhuman_dir.join(ACTIVE_USER_STATE_FILE);
-    if path.exists() {
-        std::fs::remove_file(&path)
-            .with_context(|| format!("Failed to remove active user state: {}", path.display()))?;
-        tracing::debug!(path = %path.display(), "active user cleared");
-    }
-    Ok(())
-}
-
-/// Returns the user-scoped openhuman directory for the given user id:
-/// `{default_openhuman_dir}/users/{user_id}`.
-pub fn user_openhuman_dir(default_openhuman_dir: &Path, user_id: &str) -> PathBuf {
-    default_openhuman_dir.join("users").join(user_id)
-}
-
-/// Stable id used to scope the openhuman directory before any user has
-/// logged in.  All memory, state, config, sessions and workspace files
-/// created on first init land under `{root}/users/{PRE_LOGIN_USER_ID}`
-/// so nothing is ever written directly at the root `.openhuman` path.
-///
-/// On first successful login, this directory is migrated into the real
-/// user-scoped directory (see `credentials::ops::store_session`).
-pub const PRE_LOGIN_USER_ID: &str = "local";
-
-/// Returns the pre-login (unauthenticated) user directory:
-/// `{default_openhuman_dir}/users/local`.
-pub fn pre_login_user_dir(default_openhuman_dir: &Path) -> PathBuf {
-    user_openhuman_dir(default_openhuman_dir, PRE_LOGIN_USER_ID)
-}
-
-/// Try to parse config TOML. On failure, try `.bak`, then fall back to `Config::default()`.
-///
-/// Returns `(config, was_corrupted)` where `was_corrupted == true` means the
-/// primary failed to parse and the caller is responsible for archiving the
-/// corrupt primary and persisting the recovered/default config.
-///
-/// This is a standalone async function (not a method on Config) so it can be
-/// called from both `load_or_init` and `load_from_default_paths`.
-///
-/// **Why the parse runs via `spawn_blocking`:** `toml::from_str::<Config>`
-/// is a recursive-descent parser whose serde-monomorphised `Visitor`
-/// frames for our deeply-nested `Config` cost several KB each. When this
-/// function is called from the bottom of a deep async tower — e.g.
-/// `composio_list_tools` reloading the config per call (#1710 Wave 4),
-/// reached via `chat → orchestrator → delegate_to_integrations_agent →
-/// sub-agent → composio_*` — running the parse inline on the tokio
-/// worker thread blows the ~2 MB worker stack and aborts the in-process
-/// core with `SIGBUS / KERN_PROTECTION_FAILURE` (see `crahs.log`,
-/// 2026-05-17, and `tests/composio_list_tools_stack_overflow_regression.rs`).
-/// Moving the parse onto the blocking-pool gives it a *fresh* thread
-/// stack with no async tower above it, so the same parser frames easily
-/// fit. (An earlier draft of this fix also fronted
-/// `config::ops::load_config_with_timeout` with a per-process cache to
-/// skip the parse on repeat calls, but it was reverted — the in-process
-/// integration tests in `tests/json_rpc_e2e.rs` reuse workspace paths
-/// and load config mid-mutation, racing the cache. The spawn_blocking
-/// move is sufficient on its own once paired with the Tauri worker
-/// stack bump in `app/src-tauri/src/lib.rs`.)
 async fn parse_config_with_recovery(config_path: &Path, contents: &str) -> (Config, bool) {
     let parse_err = match parse_toml_off_worker(contents.to_string()).await {
         Ok(config) => {
